+++ front/e2e/mockData/index.ts
@@ -1,0 +1,4 @@
+export * from './popularShows';
+export * from './topRatedShows';
+export * from './searchPokerFace';
+export * from './misc';
