+++ front/e2e/mockData/misc.ts
@@ -1,0 +1,5 @@
+export const email = 'playwright@test.com';
+export const password = 'password123';
+
+// 247718 is MobLand, 120998 is Poker Face
+export const followResponse = [247718, 120998];
