+++ front/e2e/tests/calendarPage.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '../fixtures';
 import { login } from '../helpers';
-import { email } from '../mockData/mockData';
+import { email } from '../mockData';
 import { mockRequest } from '../mockRequest';
 import { baseUrl } from '../playwright.config';
